Add working base executable
Base executable delegates to other commands prefixed by `ctool-`. If the
exact command name does not exist, it will search for all available
commands on the current user's PATH or system paths and perform a fuzzy
match and suggest that to the user (if there's a good enough match). A
possible option for the future is to automatically run the closest match
-- I know there have been times where I wished `git` would do that for
me.